Add segment rendering using bsp

+++ src/config.rs
@@ -3,4 +3,4 @@ pub const SCALE: u32 = 4;
 pub const WIN_RES: (u32, u32) = (DOOM_RES.0 * SCALE, DOOM_RES.1 * SCALE);
 pub const CENTER: (u32, u32) = (WIN_RES.0 / 2, WIN_RES.1 / 2);
 
-pub const SUBSECTOR_IDENTIFIER: usize = 0x8000;
+pub const SUBSECTOR_IDENTIFIER: u16 = 0x8000;
